refactor: organize components and scoped styles
