+++ src/components/ImageGallery/ImageGallery.css
@@ -1,0 +1,49 @@
+.image-gallery {
+  margin: 1rem 0;
+}
+
+.main-image {
+  width: 100%;
+  border-radius: 0.5rem;
+  margin-bottom: 0.5rem;
+  max-height: 70vh;
+  object-fit: contain;
+}
+
+.thumbnail-row {
+  display: flex;
+  gap: 0.375rem;
+  margin-top: 0.625rem;
+  margin-left: 0.5rem;
+  margin-right: 0.5rem;
+  overflow-x: auto;
+  padding: 0.375rem;
+  border: 1px solid #ccc;
+  border-radius: 0.25rem;
+}
+
+.thumbnail {
+  width: 5rem;
+  height: 3.75rem;
+  object-fit: cover;
+  border-radius: 0.25rem;
+  cursor: pointer;
+  transition: transform 0.2s, border 0.2s;
+  border: 0.125rem solid transparent;
+}
+
+.thumbnail:hover {
+  transform: scale(1.05);
+}
+
+.thumbnail.selected {
+  border: 0.125rem solid #004080;
+}
+
+@media (max-width: 768px) {
+  .thumbnail-row {
+    flex-wrap: wrap;
+    justify-content: center;
+    margin: 0.625rem 0.5rem 0;
+  }
+}
