+++ src/components/PdfViewer/PdfViewer.css
@@ -1,0 +1,10 @@
+.pdf-panel {
+  width: 100%;
+}
+
+.pdf-frame {
+  width: 100rem;
+  max-width: 100%;
+  height: 100vh;
+  border: none;
+}
