+++ src/components/SourceAttribution/SourceAttribution.css
@@ -1,0 +1,3 @@
+.source {
+  margin-top: 1rem;
+}
