+++ src/components/TabPanel/TabPanel.css
@@ -1,0 +1,141 @@
+.tabcontent {
+  display: none;
+  padding: 20vh 10rem 1.875rem;
+  width: 100%;
+  margin: auto;
+}
+
+.tabcontent.active {
+  display: flex;
+}
+
+.tab-inner {
+  display: flex;
+  width: 100%;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 2.5rem;
+  margin-top: 1.25rem;
+  max-width: 100%;
+}
+
+.content {
+  flex: 1;
+  min-width: 0;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.sidebar {
+  background-color: #f0f0f0;
+  padding: 0.75rem;
+  border-radius: 0.375rem;
+  width: min(31.25rem, 80vw);
+  max-width: 80%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  flex-shrink: 0;
+  transform: translateX(-100%);
+  transition: transform 0.3s ease;
+  z-index: 1001;
+  max-height: 100vh;
+  overflow-y: auto;
+  -webkit-overflow-scrolling: touch;
+  overscroll-behavior: contain;
+}
+
+.sidebar.open {
+  transform: translateX(0);
+}
+
+.sidebar-link {
+  display: block;
+  padding: 0.625rem 0.75rem;
+  margin-bottom: 0.5rem;
+  background-color: #fff;
+  border-left: 0.25rem solid transparent;
+  border-radius: 0.25rem;
+  font-weight: bold;
+  color: #004080;
+  text-decoration: none;
+  transition: background-color 0.2s, border-left 0.2s;
+  width: 100%;
+  text-align: left;
+}
+
+.sidebar button.sidebar-link {
+  border: none;
+  background-color: #fff;
+  cursor: pointer;
+  font: inherit;
+}
+
+.sidebar-link:hover,
+.sidebar-link.active-link {
+  background-color: #e6f0ff;
+  border-left: 0.25rem solid #004080;
+}
+
+.performance-item,
+.video-item {
+  background-color: white;
+  border-radius: 0.5rem;
+  padding: 1.25rem;
+  box-shadow: 0 0 0.625rem rgba(0, 0, 0, 0.05);
+  margin-bottom: 1.875rem;
+  scroll-margin-top: 20vh;
+}
+
+.performance-item {
+  padding: 0;
+  margin: 0 0 1rem;
+  box-shadow: none;
+  flex: 1;
+}
+
+.performance-item h2 {
+  margin-top: 0;
+}
+
+.performance-item video,
+.performance-item img,
+.performance-item iframe {
+  width: 100%;
+  border-radius: 0.5rem;
+  margin-top: 0.625rem;
+}
+
+.text-content {
+  margin: 1rem 0;
+}
+
+@media (max-width: 768px) {
+  .sidebar {
+    width: 80vw;
+  }
+
+  .tabcontent {
+    padding: 20vh 2rem 1.875rem;
+  }
+
+  .tab-inner {
+    flex-direction: column;
+  }
+
+  .performance-item {
+    padding: 0.625rem;
+  }
+
+  .performance-item h2 {
+    font-size: 1.125rem;
+  }
+
+  .performance-item iframe {
+    width: 100% !important;
+    height: auto;
+  }
+}
